requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
boto3==1.37.7
botocore==1.37.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-database-url==2.3.0
Django==4.2.19
django-allauth==65.4.1
django-bootstrap-v5==1.0.11
django-heroku==0.3.1
django-storages==1.14.5
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==1.26.20
whitenoise==6.9.0
